Refactor filter container
