+++ src/components/ConditionalFilter/useFilterContainer.ts
@@ -1,115 +1,36 @@
-// @ts-strict-ignore
-import { useState } from "react";
-
-import { FilterElement } from "./FilterElement";
+import { FilterContainer } from "./FilterElement";
 import { ConditionValue, ItemOption } from "./FilterElement/ConditionValue";
-
-export type FilterContainer = Array<string | FilterElement | FilterContainer>;
+import { useContainerState } from "./useContainerState";
 
 export const useFilterContainer = (initialValue: FilterContainer) => {
-  const [value, setValue] = useState(initialValue);
+  const { value, updateAt, removeAt, createEmpty } = useContainerState(initialValue)
 
   const addEmpty = () => {
-    const newValue = [];
-    if (value.length > 0) {
-      newValue.push("OR");
-    }
-
-    newValue.push(FilterElement.createEmpty());
-
-    setValue(v => v.concat(newValue));
-  };
-
-  const removeAt = (position: string) => {
-    const index = parseInt(position, 10);
-
-    if (value.length > 0) {
-      setValue(v =>
-        v.filter((_, elIndex) => ![index - 1, index].includes(elIndex)),
-      );
-      return;
-    }
-
-    setValue(v => v.filter((_, elIndex) => ![index].includes(elIndex)));
+    createEmpty()
   };
 
   const updateLeftOperator = (position: string, leftOperator: any) => {
-    const index = parseInt(position, 10);
-    setValue(v =>
-      v.map((el, elIndex) => {
-        if (elIndex === index && typeof el !== "string" && !Array.isArray(el)) {
-          el.updateLeftOperator(leftOperator);
-        }
-
-        return el;
-      }),
-    );
+    updateAt(position, (el) => el.updateLeftOperator(leftOperator))
   };
 
   const updateLeftLoadingState = (position: string, loading: boolean) => {
-    const index = parseInt(position, 10);
-    setValue(v =>
-      v.map((el, elIndex) => {
-        if (elIndex === index && typeof el !== "string" && !Array.isArray(el)) {
-          el.updateLeftLoadingState(loading);
-        }
-
-        return el;
-      }),
-    );
+    updateAt(position, (el) => el.updateLeftLoadingState(loading))
   };
 
   const updateRightOperator = (position: string, rightOperator: ConditionValue) => {
-    const index = parseInt(position, 10);
-    setValue(v =>
-      v.map((el, elIndex) => {
-        if (elIndex === index && typeof el !== "string" && !Array.isArray(el)) {
-          el.updateRightOperator(rightOperator);
-        }
-
-        return el;
-      }),
-    );
+    updateAt(position, (el) => el.updateRightOperator(rightOperator))
   };
 
   const updateRightOptions = (position: string, options: ItemOption[]) => {
-    const index = parseInt(position, 10);
-    setValue(v =>
-      v.map((el, elIndex) => {
-        if (elIndex === index && typeof el !== "string" && !Array.isArray(el)) {
-          el.updateRightOptions(options);
-        }
-
-        return el;
-      }),
-    );
+    updateAt(position, (el) => el.updateRightOptions(options))
   };
 
   const updateRightLoadingState = (position: string, loading: boolean) => {
-    const index = parseInt(position, 10);
-    setValue(v =>
-      v.map((el, elIndex) => {
-        if (elIndex === index && typeof el !== "string" && !Array.isArray(el)) {
-          el.updateRightLoadingState(loading);
-        }
-
-        return el;
-      }),
-    );
+    updateAt(position, (el) => el.updateRightLoadingState(loading))
   };
 
   const updateCondition = (position: string, conditionValue: any) => {
-    const index = parseInt(position, 10);
-
-    setValue(v =>
-      v.map((el, elIndex) => {
-        if (elIndex === index && typeof el !== "string" && !Array.isArray(el)) {
-          el.updateCondition(conditionValue);
-        }
-
-        return el;
-      }),
-    );
+    updateAt(position, (el) => el.updateCondition(conditionValue))
   };
 
   return {
